+++ lib/src/rules/avoid_null_checks_in_equality_operators.dart
